refactor: split version API module by responsibility
