+++ frontend/src/api/version/hooks.ts
@@ -1,0 +1,15 @@
+import { useQuery } from '@tanstack/react-query';
+import { fetchAppVersion } from '@/api/version/requests';
+import { appVersionKeys } from '@/api/queryKeys';
+
+/**
+ * Polls the backend version endpoint for app update notices
+ */
+export function useAppVersion() {
+  return useQuery({
+    queryKey: appVersionKeys.version(),
+    queryFn: fetchAppVersion,
+    staleTime: 15 * 60 * 1000,
+    gcTime: Infinity,
+  });
+}
